pracexam1

+++ function-1-1.cpp
@@ -1,5 +1,5 @@
 double arrayMin(double* array, int size){
-    int min = array[0];
+    double min = array[0];
     for( int i  = 0; i< size; i++){
         if(array[i]<min){
         min  = array[i];
